--- app/lib/malware_scanning.rb
@@ -1,22 +1,0 @@
-module MalwareScanning
-  private
-
-    def malware_scan_result(original_file)
-      MalwareScanner.call(
-        file_path: original_file.tempfile.path,
-        file_details: {
-          size: original_file_size(original_file),
-          name: original_file.original_filename,
-          content_type: original_file.content_type,
-        },
-      )
-    end
-
-    def original_file_error_for(error_type, options = {})
-      I18n.t("activemodel.errors.models.#{error_path}.attributes.original_file.#{error_type}", **options)
-    end
-
-    def original_file_size(original_file)
-      File.size(original_file.tempfile)
-    end
-  end
