+++ app/models/malware_scan_result.rb
@@ -1,3 +1,3 @@
 class MalwareScanResult < ApplicationRecord
-    belongs_to :uploader, polymorphic: true
+#   belongs_to :uploader, polymorphic: true
 end
