+++ app/services/malware_scanner.rb
@@ -8,16 +8,16 @@ class MalwareScanner
   LOCAL_COMMAND     = %w[clamdscan --fdpass --no-summary].freeze
   CONTAINER_COMMAND = %w[clamdscan -c config/clamd.container.conf --stream --no-summary].freeze
 
-  def initialize(file_path:, uploader: nil, file_details: nil, save_result: true)
+  def initialize(file_path:, file_details: nil, save_result: true)
     @file_path    = file_path
-    @uploader     = uploader
+    # @uploader     = uploader
     @file_details = file_details
     @save_result  = save_result
   end
 
   def call
     result = MalwareScanResult.new(
-      uploader:,
+      # uploader:,
       virus_found: virus_found?,
       scan_result: scan_result.result,
       scanner_working: scanner_working?,
@@ -31,7 +31,7 @@ class MalwareScanner
 
 private
 
-  attr_reader :file_path, :uploader, :file_details, :save_result
+  attr_reader :file_path, :file_details, :uploader, :save_result
 
   def virus_found?
     @virus_found ||= !scan_result.status.success?
@@ -40,7 +40,7 @@ private
   def scanner_working?
     return true unless scan_result.status.exitstatus == 2
 
-    AlertManager.capture_message("ClamdscanError, #{scan_result.error.presence || scan_result.result}")
+    Sentry.capture_message("ClamdscanError, #{scan_result.error.presence || scan_result.result}")
     false
   end
