+++ config/clamd.container.conf
@@ -1,0 +1,2 @@
+TCPSocket 3310
+TCPAddr localhost
