+++ db/migrate/20230607074210_create_malware_scan_results.rb
@@ -1,0 +1,11 @@
+class CreateMalwareScanResults < ActiveRecord::Migration[7.0]
+  def change
+    create_table :malware_scan_results, id: :uuid do |t|
+      t.boolean :virus_found, null: false
+      t.text :scan_result
+      t.json :file_details
+      t.boolean :scanner_working
+      t.timestamps
+    end
+  end
+end
